Move marking functions into new module
